logics are separated with custom hooks
