+++ want_to_do/src/components/Home/EditTodoForm.tsx
@@ -1,36 +1,22 @@
-import {useUpdateTodo} from "../../hooks/todos";
-import {useEffect, useState} from "react";
-import {useRecoilState} from "recoil";
-import {todoToUpdateState} from "../../states/todo/todoToUpdateState";
 import TodoForm from "./TodoForm";
+import useEditTodoForm from "./hooks/useEditTodoForm";
 
 const EditTodoForm = () => {
-  const [todoToUpdate, setTodoToUpdate] = useRecoilState(todoToUpdateState);
-  if (!todoToUpdate) { return null; }
-
-  const [title, setTitle] = useState("");
-  const [content, setContent] = useState("");
-  const updateTodo = useUpdateTodo();
-
-  useEffect(() => {
-    setTitle(todoToUpdate.title);
-    setContent(todoToUpdate.content);
-  },[todoToUpdate]);
-
-  const handleUpdateTodo = () => {
-    if (title === "") {
-      alert("할 일을 입력해주세요.");
-      return;
-    }
-    updateTodo.mutate({id: todoToUpdate.id, requestBody: {title, content}});
-  }
+  const {
+    title,
+    content,
+    setTitle,
+    setContent,
+    handleUpdateTodo,
+    handleCancel
+  } = useEditTodoForm();
 
   return (
     <TodoForm title={title}
               content={content}
               setTitle={setTitle}
               setContent={setContent}
-              handleCancel={() => {setTodoToUpdate(null)}}
+              handleCancel={handleCancel}
               handleSubmit={handleUpdateTodo}
               cancelLabel={"취소"}
               submitLabel={"수정"}
