+++ want_to_do/src/components/Home/TodoListCell.tsx
@@ -4,38 +4,21 @@ import {useRecoilState, useRecoilValue} from "recoil";
 import {isEditModeOnState} from "../../states/todo/isEditModeOnState";
 import {todoToUpdateState} from "../../states/todo/todoToUpdateState";
 import {useNavigate} from "react-router-dom";
+import useTodoListCell from "./hooks/useTodoListCell";
 
 interface TodoListCellProps {
   todo: Todo;
 }
 
-const previewContent = (content: string) => {
-  if (content.length > 8) {
-    return content.slice(0, 10) + "...";
-  }
-  return content;
-}
-
 export const TodoListCell = ({todo}: TodoListCellProps) => {
 
-  const isEditModeOn = useRecoilValue(isEditModeOnState);
-  const [todoToUpdate, setTodoToUpdate] = useRecoilState(todoToUpdateState);
-  const deleteTodo = useDeleteTodo();
-  const navigate = useNavigate();
-
-  const showDetailedContent = () => {
-    navigate(`/home/${todo.id}`);
-  }
-
-  const handleEditTodo = (event: React.MouseEvent) => {
-    event.stopPropagation()
-    todoToUpdate ? setTodoToUpdate(null) : setTodoToUpdate(todo);
-  }
-
-  const handleDeleteTodo = (event: React.MouseEvent) => {
-    event.stopPropagation()
-    confirm("정말 삭제하시겠습니까?") && deleteTodo.mutate(todo.id);
-  }
+  const {
+    handleEditTodo,
+    handleDeleteTodo,
+    showDetailedContent,
+    isEditModeOn,
+    previewContent
+  } = useTodoListCell(todo);
 
   const EditTodoButtons = () => {
     return (
